Control.Parallel.Meta

Description

Documentation

data Startup

The Startup component of a Resource is a callback that implements initialisation behaviour. For example, it might contact remote hosts, spawn threads, or initialise hardware such as GPUs.

Constructors

Startup

Fields

runStartup :: Gang -> IO ()

data WorkSearch

The WorkSearch component of a Resource is a callback that responds to requests for work from meta-workers. The arguments to WorkSearch are the scheduler state for the current thread and a reference to all workers in the program.

Constructors

WorkSearch

Fields

runWorkSearch :: Worker -> IO (Maybe Range)

data Resource

A Resource provides an abstraction of heterogeneous execution resources that may be combined. Composition of resources is left-biased. That is, if if resource1 always returns work from its WorkSearch, then the composed resource resource1 <> resource2 will never request work from resource2.

Constructors

Resource

Fields

startup :: Startup

workSearch :: WorkSearch

type Action = Int -> Int -> Int -> IO ()

An action to execute. The first parameters are the start and end indices of the range this action should process, and the final is the ID of the thread doing the work.

data Executable

An Executable provides a callback that can be used to run a provided function using an encapsulated work-stealing gang of threads.

Constructors

Executable

Fields

runExecutable :: Int -> Range -> Finalise -> Maybe Action -> Action -> IO ()

data Finalise

The Finalise component of an executable is an action the thread applies after processing the work function, given its thread id the ranges that this thread actually handled.

Constructors

Finalise

Fields

runFinalise :: Int -> Seq Range -> IO ()

runParIO :: Resource -> Gang -> Range -> Maybe Action -> Action -> Finalise -> IO ()

Run a parallel work-stealing operation.

Each thread initialises its work queue with an equally sized chunk of the work. Threads keep working until their work search returns Nothing, at which point the thread exits. In our LBS implementation, a worker thread takes a small chunk of its work range to process, and places the remainder back onto its deque. Thus the work queues are only empty:

(a) Briefly during the scheduling process; or

(b) After the deque has been robbed. If the stolen chunk is large enough, the stealee will split the remainder onto its deque to be stolen; or

(c) There is no more work.

As long as the thread makes a small number of retries, this should correctly balance the work without too much scheduler overhead.

An alternative to every thread initialising with an even chunk size is to put the entire range onto the first worker and then have the scheduler handle the decomposition. However, since we are playing fast-and-loose with the exit condition, this is a little racy.

TLM NOTE:

Should threads check whether the work queues of all threads are empty before deciding to exit? If the PPT is too large then threads might not react quickly enough to splitting once their deque is emptied. Maybe the first thread to return Nothing can probe the queues to see if they are all empty. If True, write into a shared MVar to signal to the others that it is time to exit. But, that still assumes that the PPT is not so large that the queues are always empty.

TLM TODO:

Splitting work should probably be biased to cache-size chunks, rather than trying to split exactly evenly.

seqIO :: Resource -> Gang -> Range -> Maybe Action -> Action -> Finalise -> IO ()

parIO :: Resource -> Gang -> Range -> Maybe Action -> Action -> Finalise -> IO ()