Control.Parallel.Meta.Resource.Backoff

Description

This module implements exponential backoff so as to prevent spamming of stealing actions. Most scheduler compositions should include this at the bottom of the stack.

Inspired by the meta-par package. This package has a BSD license. http://hackage.haskell.org/package/meta-par

Documentation

mkResource :: Resource

defaultWorkSearch :: WorkSearch

mkWorkSearch :: Int -> Int -> WorkSearch

To construct the work search, we need to know the minimum and maximum amount of time, in nanoseconds, to sleep. The exponential backoff policy is always the same: it starts at 1µs and doubles at every failure.

The thing that changes over time is whether sleeping actually occurs. For example, the defaultWorkSearch:

mkWorkSearch 100 10000

will not sleep for the first 7 invocations (until 128), and then will sleep an amount that doubles each time until it surpasses the maximum, at which point it will always sleep for the maximum time (10ms)