Control.Parallel.Meta.Resource.SMP

Description

This module implements a resource for SMP parallelism. It is suitable as a base for combining a bunch of resources that can steal cheaply from each other.

Inspired by the meta-par package. This package has a BSD license. http://hackage.haskell.org/package/meta-par

Documentation

mkResource

Arguments

:: Int
number of steal attempts per WorkSearch

-> Gang
the workers that will steal amongst each other

-> Resource

Create an SMP (symmetric multiprocessor) resource where all underlying workers have uniform access to shared resources such as memory. Thus workers at this level can steal cheaply from each other.

mkWorkSearch :: Int -> Gang -> WorkSearch

Given a set of workers and a number of steal attempts per worker, return a work search function. Steals from workers in this gang are considered cheap and uniform.

Note: [Number of retries in SMP resource]

A large number of retries in the work search will prevent the search function from traversing the resource stack. This can result in spamming of stealing actions. In particular, the exponential backoff resource usually placed at the bottom of the stack may never be reached. Thus a balance is required between number of times to traverse each level and number of times to traverse the entire resource stack.