Control.Parallel.Meta.Trans.LBS

Description

This module implements a lazy binary splitting resource transformer. It is suitable to add an adaptive runtime work-stealing component to resource.

Documentation

mkResource

Arguments

:: Int
profitable parallelism threshold

-> Resource
the resource to modify

-> Resource

Transform the WorkSearch function of the given Resource to include a lazy binary splitting work stealing scheduler.

mkWorkSearch

Arguments

:: Int
profitable parallelism threshold

-> WorkSearch
the basic work search method to modify

-> WorkSearch

This transforms the WorkSearch function to add a lazy binary splitting operation on top of an existing (work-stealing) scheduler. On receiving a unit of work, the scheduler proceeds as follows:

If the number of iterations is less than the profitable parallelism threshold (ppt), execute the remaining iterations and exit, else (2).

Check this worker's remaining work queue. If it is not empty, then execute ppt iterations, then go to (1) with the remainder.

If the remaining work queue is empty, split this work chunk in half. Place the second half onto the remaining work queue and go to (1).