Control.Parallel.Meta.Worker

Description

Documentation

type Gang = Vector Worker

The Gang structure tracks the state of all workers in the program. It starts empty, and workers append to it as they are brought online. Although the vector append operation is expensive, it is expected it is only called occasionally; e.g. at program initialisation. So, we prioritise for constant lookup of the worker structure, which will be done frequently during the work search.

data Worker

The Worker is the per-worker-thread state.

If the worker has work that can be stolen by other processors, it is stored in the workpool. Thieves treat the workpool as a stack which can be popped on the right, where as the owner can both push and pop on the left.

In the lazy binary splitting work stealing setup, a worker processes its range in chunks, checking the state of its workpool periodically. Whenever the queue is empty, it splits it's current workload in two so that the second half can be stolen by another processor.

Constructors

Worker

Fields

workerId :: !Int

requestVar :: !(MVar Req)

resultVar :: !(MVar ())

workpool :: !(WSDeque Range)

consecutiveFailures :: !(IORef Int)

rngState :: !GenIO

data Req

The Req type encapsulates work requests for individual workers

Constructors

ReqDo (Int -> IO ())
Instruct the worker to run the given action

ReqShutdown
Tell the worker to exit. The worker should signal that it received the request by writing its result var before exiting.

gangSize :: Gang -> Int

O(1). Yield the number of threads in the Gang.

forkGang :: Int -> IO Gang

Create a set of workers. This is a somewhat expensive function, so it is expected that it is called only occasionally (e.g. once per program execution).

forkGangOn :: [Int] -> IO Gang

Create a set of workers on specific capabilities. Note that the thread ID passed to the gangWorker is the index of this worker in the gang structure, not necessarily the capability is is spawned on.

gangIO :: Gang -> (Int -> IO ()) -> IO ()

Issue work requests to the gang and wait until they complete

exhausted :: Gang -> IO Bool

Check whether the work queues of the gang are all empty