Data.Array.Accelerate.LLVM.Array.Data

Description

Documentation

class Async arch => Remote arch where

Minimal complete definition

Nothing

Methods

allocateRemote :: (Shape sh, Elt e) => sh -> LLVM arch (Array sh e)

Allocate a new uninitialised array on the remote device.

useRemoteR

Arguments

:: (ArrayElt e, ArrayPtrs e ~ Ptr a, Storable a, Typeable a, Typeable e)

=> Int
number of elements in the payload

-> Maybe (StreamR arch)
execute synchronously w.r.t. this execution stream

-> ArrayData e
array payload

-> LLVM arch ()

Use the given immutable array on the remote device. Since the source array is immutable, the allocator can evict and re-upload the data as necessary without copy-back.

copyToRemoteR

Arguments

:: (ArrayElt e, ArrayPtrs e ~ Ptr a, Storable a, Typeable a, Typeable e)

=> Int
index of first element to copy

-> Int

-> Maybe (StreamR arch)
execute synchronously w.r.t. this execution stream

-> ArrayData e
array payload

-> LLVM arch ()

Upload a section of an array from the host to the remote device. Only the elements between the given indices (inclusive left, exclusive right) are transferred.

copyToHostR

Arguments

:: (ArrayElt e, ArrayPtrs e ~ Ptr a, Storable a, Typeable a, Typeable e)

=> Int
index of the first element to copy

-> Int

-> Maybe (StreamR arch)
execute synchronously w.r.t. this execution stream

-> ArrayData e
array payload

-> LLVM arch ()

Copy a section of an array from the remote device back to the host. The elements between the given indices (inclusive left, exclusive right) are transferred.,v

copyToPeerR

Arguments

:: (ArrayElt e, ArrayPtrs e ~ Ptr a, Storable a, Typeable a, Typeable e)

=> Int
index of the first element to copy

-> Int

-> arch
remote device to copy to

-> Maybe (StreamR arch)
execute synchronously w.r.t. this execution stream

-> ArrayData e
array payload

-> LLVM arch ()

Copy a section of an array between two remote instances of the same type. This may be more efficient than copying to the host and then to the second remote instance (e.g. DMA between two CUDA devices). The elements between the given indices (inclusive left, exclusive right) are transferred.

indexRemote :: Array sh e -> Int -> LLVM arch e

Read a single element from the array at a given row-major index

newRemote :: (Remote arch, Shape sh, Elt e) => sh -> (sh -> e) -> LLVM arch (Array sh e)

Create a new array from its representation on the host, and upload it to a new remote array.

useRemote :: (Remote arch, Arrays arrs) => arrs -> LLVM arch arrs

Upload an immutable array from the host to the remote device. This is a synchronous operation in that it will not return until the transfer completes, but the individual array payloads will be uploaded concurrently if possible.

useRemoteAsync :: (Remote arch, Arrays arrs) => arrs -> StreamR arch -> LLVM arch (AsyncR arch arrs)

Upload an immutable array from the host to the remote device, asynchronously. This will upload each array payload in a separate execution stream, thereby making us of multiple memcpy engines (where available).

copyToRemote :: (Remote arch, Arrays a) => a -> LLVM arch a

Uploading existing arrays from the host to the remote device. This is synchronous with respect to the calling thread, but the individual array payloads may themselves be transferred concurrently.

copyToRemoteAsync :: (Remote arch, Arrays a) => a -> StreamR arch -> LLVM arch (AsyncR arch a)

Upload an existing array to the remote device, asynchronously.

copyToHost :: (Remote arch, Arrays a) => a -> LLVM arch a

Copy an array from the remote device to the host. This is synchronous with respect to the calling thread, but the individual array payloads may themselves be transferred concurrently.

copyToHostAsync :: (Remote arch, Arrays a) => a -> StreamR arch -> LLVM arch (AsyncR arch a)

Copy an array from the remote device to the host, asynchronously

copyToPeer :: (Remote arch, Arrays a) => arch -> a -> LLVM arch a

Copy arrays between two remote instances of the same type. This may be more efficient than copying to the host and then to the second remote instance (e.g. DMA between CUDA devices).

copyToPeerAsync :: (Remote arch, Arrays a) => arch -> a -> StreamR arch -> LLVM arch (AsyncR arch a)

As copyToPeer, asynchronously.

runIndexArray :: forall m sh e. Monad m => (forall e a. (ArrayElt e, ArrayPtrs e ~ Ptr a, Storable a, Typeable a, Typeable e) => ArrayData e -> Int -> m a) -> Array sh e -> Int -> m e

Read a single element from an array at the given row-major index.

runArrays :: forall m arrs. (Functor m, Applicative m, Monad m, Arrays arrs) => arrs -> (forall sh e. Array sh e -> m (Array sh e)) -> m arrs

Generalised function to traverse the Arrays structure

runArray :: forall m sh e. (Functor m, Applicative m, Monad m) => Array sh e -> (forall e' p. (ArrayElt e', ArrayPtrs e' ~ Ptr p, Storable p, Typeable p, Typeable e') => ArrayData e' -> m (ArrayData e')) -> m (Array sh e)

Generalised function to traverse the ArrayData structure with one additional argument

module Data.Array.Accelerate.Array.Data