Data.Array.Accelerate.LLVM.Array.Nursery

Description

Documentation

data Nursery a

The nursery is primarily designed as a place to store device memory arrays that are no longer needed. Often it is quicker to reuse an existing array, rather than calling into the device API in order to allocate a fresh array.

The nursery is wrapped in an MVar so that several threads may safely access it concurrently.

Constructors

Nursery !(NRS a) !(Weak (NRS a))

type NRS a = MVar (IntMap (Seq a))

new :: forall a. (a -> IO ()) -> IO (Nursery a)

Create a fresh nursery

When the nursery is garbage collected, the provided function will be run on each element of the map. In the context of array storage, this would be to free the stashed memory.

lookup :: Int -> Nursery a -> IO (Maybe a)

Look up an entry in the nursery with a given key.

insert :: Int -> a -> NRS a -> IO ()

Add an entry into the nursery

cleanup :: forall a. (a -> IO ()) -> NRS a -> IO ()

Delete all entries from the nursery.

trace :: String -> IO a -> IO a

message :: String -> IO ()