Data.Array.Accelerate.LLVM.Array.Table

Description

Documentation

data MemoryTable c

The memory table is used to associate a host-side Accelerate array with a corresponding array in on a remote target.

Old entries in the table are garbage collected from the table once the host array is no longer reachable on the heap. The table stores weak references to its entries. Once the key becomes unreachable, a finaliser will fire and remove this entry from the table, and further attempts to dereference the weak pointer will fail.

PRECONDITION: The underlying Accelerate array is pinned.

type MallocRemote c a = Int -> IO (c a)

type FreeRemote c = forall a. c a -> IO ()

new :: FreeRemote c -> IO (MemoryTable c)

Create a new memory table from host to remote arrays. When the structure is collected it will finalise all entries in the table.

member :: (ArrayElt e, ArrayPtrs e ~ Ptr a) => MemoryTable c -> ArrayData e -> IO Bool

Is the key a _valid_ member of the memory table?

Because of the possibility of collisions in the host key, this also ensures that the entries still refers to valid data. If not, we run the finaliser, which also deallocates the remote data and removes its entry from the table. Note that it is important for the main thread to yield at that point so that the finaliser can run immediately.

See: [Host arrays as memory table keys]

lookup :: (ArrayElt e, ArrayPtrs e ~ Ptr a, Typeable b) => MemoryTable c -> ArrayData e -> IO (Maybe (c b))

Lookup the remote array corresponding to the given host-side array. This assumes that all key-value pairs in the memory table refer to valid data. If you are not sure, or want to check whether there is a remote array for the given key, use member instead.

See note: [Host arrays as memory table keys].

malloc :: forall c e a. (ArrayElt e, ArrayPtrs e ~ Ptr a, Typeable a, Storable a) => FreeRemote c -> MallocRemote c a -> MemoryTable c -> ArrayData e -> Int -> IO (c a)

Allocate a new remote array and associate it with the given host side array. This will attempt to reuse an old array from the nursery if possible.

In order to increase the hit rate to the nursery, allocations are not made for the exact number of bytes requested, but instead rounded up to a chunk/page size.

cleanup :: FreeRemote c -> MemoryTable c -> IO ()

Cleanup the memory table in an attempt to create more free space on the remote device. This removes all entries from the nursery, and finalises any unreachable arrays.