Data.Array.Accelerate.LLVM.CodeGen

Description

Documentation

class Skeleton arch where

A class covering code generation for all of the primitive array operations. Client backends implement an instance of this class.

Minimal complete definition: * generate * fold, fold1, foldSeg, fold1Seg * scanl, scanl', scanl1, scanr, scanr', scanr1 * permute

Minimal complete definition

generate, fold, fold1, foldSeg, fold1Seg, scanl, scanl', scanl1, scanr, scanr', scanr1, permute

Methods

generate :: (Shape sh, Elt e) => arch -> Gamma aenv -> IRFun1 arch aenv (sh -> e) -> CodeGen (IROpenAcc arch aenv (Array sh e))

transform :: (Shape sh, Shape sh', Elt a, Elt b) => arch -> Gamma aenv -> IRFun1 arch aenv (sh' -> sh) -> IRFun1 arch aenv (a -> b) -> IRDelayed arch aenv (Array sh a) -> CodeGen (IROpenAcc arch aenv (Array sh' b))

map :: (Shape sh, Elt a, Elt b) => arch -> Gamma aenv -> IRFun1 arch aenv (a -> b) -> IRDelayed arch aenv (Array sh a) -> CodeGen (IROpenAcc arch aenv (Array sh b))

fold :: (Shape sh, Elt e) => arch -> Gamma aenv -> IRFun2 arch aenv (e -> e -> e) -> IRExp arch aenv e -> IRDelayed arch aenv (Array (sh :. Int) e) -> CodeGen (IROpenAcc arch aenv (Array sh e))

fold1 :: (Shape sh, Elt e) => arch -> Gamma aenv -> IRFun2 arch aenv (e -> e -> e) -> IRDelayed arch aenv (Array (sh :. Int) e) -> CodeGen (IROpenAcc arch aenv (Array sh e))

foldSeg :: (Shape sh, Elt e, Elt i, IsIntegral i) => arch -> Gamma aenv -> IRFun2 arch aenv (e -> e -> e) -> IRExp arch aenv e -> IRDelayed arch aenv (Array (sh :. Int) e) -> IRDelayed arch aenv (Segments i) -> CodeGen (IROpenAcc arch aenv (Array (sh :. Int) e))

fold1Seg :: (Shape sh, Elt e, Elt i, IsIntegral i) => arch -> Gamma aenv -> IRFun2 arch aenv (e -> e -> e) -> IRDelayed arch aenv (Array (sh :. Int) e) -> IRDelayed arch aenv (Segments i) -> CodeGen (IROpenAcc arch aenv (Array (sh :. Int) e))

scanl :: Elt e => arch -> Gamma aenv -> IRFun2 arch aenv (e -> e -> e) -> IRExp arch aenv e -> IRDelayed arch aenv (Vector e) -> CodeGen (IROpenAcc arch aenv (Vector e))

scanl' :: Elt e => arch -> Gamma aenv -> IRFun2 arch aenv (e -> e -> e) -> IRExp arch aenv e -> IRDelayed arch aenv (Vector e) -> CodeGen (IROpenAcc arch aenv (Vector e, Scalar e))

scanl1 :: Elt e => arch -> Gamma aenv -> IRFun2 arch aenv (e -> e -> e) -> IRDelayed arch aenv (Vector e) -> CodeGen (IROpenAcc arch aenv (Vector e))

scanr :: Elt e => arch -> Gamma aenv -> IRFun2 arch aenv (e -> e -> e) -> IRExp arch aenv e -> IRDelayed arch aenv (Vector e) -> CodeGen (IROpenAcc arch aenv (Vector e))

scanr' :: Elt e => arch -> Gamma aenv -> IRFun2 arch aenv (e -> e -> e) -> IRExp arch aenv e -> IRDelayed arch aenv (Vector e) -> CodeGen (IROpenAcc arch aenv (Vector e, Scalar e))

scanr1 :: Elt e => arch -> Gamma aenv -> IRFun2 arch aenv (e -> e -> e) -> IRDelayed arch aenv (Vector e) -> CodeGen (IROpenAcc arch aenv (Vector e))

permute :: (Shape sh, Shape sh', Elt e) => arch -> Gamma aenv -> IRFun2 arch aenv (e -> e -> e) -> IRFun1 arch aenv (sh -> sh') -> IRDelayed arch aenv (Array sh e) -> CodeGen (IROpenAcc arch aenv (Array sh' e))

backpermute :: (Shape sh, Shape sh', Elt e) => arch -> Gamma aenv -> IRFun1 arch aenv (sh' -> sh) -> IRDelayed arch aenv (Array sh e) -> CodeGen (IROpenAcc arch aenv (Array sh' e))

stencil :: (Elt a, Elt b, Stencil sh a stencil) => arch -> Gamma aenv -> IRFun1 arch aenv (stencil -> b) -> Boundary (IR a) -> IRManifest arch aenv (Array sh a) -> CodeGen (IROpenAcc arch aenv (Array sh b))

stencil2 :: (Elt a, Elt b, Elt c, Stencil sh a stencil1, Stencil sh b stencil2) => arch -> Gamma aenv -> IRFun2 arch aenv (stencil1 -> stencil2 -> c) -> Boundary (IR a) -> IRManifest arch aenv (Array sh a) -> Boundary (IR b) -> IRManifest arch aenv (Array sh b) -> CodeGen (IROpenAcc arch aenv (Array sh c))

class Expression arch where

A class covering code generation for a subset of the scalar operations. All operations (except Foreign) have a default instance, but this allows a backend to specialise the implementation for the more complex operations.

Minimal complete definition

Nothing

Methods

eforeign :: (Foreign f, Elt x, Elt y) => arch -> f x y -> IRFun1 arch () (x -> y) -> IROpenExp arch env aenv x -> IROpenExp arch env aenv y

while :: Elt a => IROpenFun1 arch env aenv (a -> Bool) -> IROpenFun1 arch env aenv (a -> a) -> IROpenExp arch env aenv a -> IROpenExp arch env aenv a

class Intrinsic arch where

During code generation we need to know the name of functions implementing certain intrinsic maths operations. Depending on the backend, these functions may not be implemented using the standard C math library.

This class allows a backend to provide a mapping from the C math library function name to the name of the function which should be called instead. The default implementation maps to the llvm intrinsic. For example:

sqrtf -> llvm.sqrt.f32 sqrt -> llvm.sqrt.f64

Minimal complete definition

Nothing

Methods

intrinsicForTarget :: arch -> HashMap String Label

llvmOfOpenAcc :: forall arch aenv arrs. (Target arch, Skeleton arch, Intrinsic arch, Expression arch) => arch -> DelayedOpenAcc aenv arrs -> Gamma aenv -> Module arch aenv arrs

Generate code for a given target architecture.