Data.Array.Accelerate.LLVM.CodeGen.Array

Description

Documentation

readArray :: forall sh e. IRArray (Array sh e) -> IR Int -> CodeGen (IR e)

Read a value from an array at the given index

writeArray :: forall sh e. IRArray (Array sh e) -> IR Int -> IR e -> CodeGen ()

Write a value into an array at the given index