Data.Array.Accelerate.LLVM.CodeGen.Base

Description

Documentation

data Name a

Objects of various sorts in LLVM IR are identified by address in the LLVM C++ API, and may be given a string name. When printed to (resp. read from) human-readable LLVM assembly, objects without string names are numbered sequentially (resp. must be numbered sequentially). String names may be quoted, and are quoted when printed if they would otherwise be misread - e.g. when containing special characters.

7

means the seventh unnamed object, while

"7"

means the object named with the string "7".

This libraries handling of UnNames during translation of the AST down into C++ IR is somewhat more forgiving than the LLVM assembly parser: it does not require that unnamed values be numbered sequentially; however, the numbers of UnNames passed into C++ cannot be preserved in the C++ objects. If the C++ IR is printed as assembly or translated into a Haskell AST, unnamed nodes will be renumbered sequentially. Thus unnamed node numbers should be thought of as having any scope limited to the Module in which they are used.

Constructors

Name String
a string name

UnName Word
a number for a nameless thing

local :: ScalarType a -> Name a -> IR a

global :: ScalarType a -> Name a -> IR a

irArray :: forall sh e. (Shape sh, Elt e) => Name (Array sh e) -> IRArray (Array sh e)

Names of array data elements

mutableArray :: forall sh e. (Shape sh, Elt e) => Name (Array sh e) -> (IRArray (Array sh e), [Parameter])

Generate typed local names for array data components as well as function parameters to bind those names

call :: GlobalFunction args t -> [FunctionAttribute] -> CodeGen (IR t)

Call a global function. The function declaration is inserted into the symbol table.

scalarParameter :: ScalarType t -> Name t -> Parameter

ptrParameter :: ScalarType t -> Name t -> Parameter

envParam :: forall aenv. Gamma aenv -> [Parameter]

Unpack the array environment into a set of input parameters to a function. The environment here refers only to the actual free array variables that are accessed by the function.

arrayParam :: forall sh e. (Shape sh, Elt e) => Name (Array sh e) -> [Parameter]

Generate function parameters for an Array with given base name.