Data.Array.Accelerate.LLVM.CodeGen.Environment

Description

Documentation

data Val env where

An environment for local scalar expression bindings, encoded at the value level as a heterogenous snoc list, and on the type level as nested tuples.

Constructors

Empty :: Val ()

Push :: Val env -> IR t -> Val (env, t)

prj :: Idx env t -> Val env -> IR t

Projection of a value from the valuation environment using a de Bruijn index.

type Gamma aenv = IntMap (Label, Idx' aenv)

A mapping between the environment index of a free array variable and the Name of that array to be used in the generated code.

This simply compresses the array indices into a continuous range, rather than directly using the integer equivalent of the de Bruijn index. Thus, the result is still sensitive to the order of let bindings, but not of any intermediate (unused) free array variables.

data Idx' aenv where

Constructors

Idx' :: (Shape sh, Elt e) => Idx aenv (Array sh e) -> Idx' aenv

aprj :: Idx aenv t -> Gamma aenv -> Name t

makeGamma :: IntMap (Idx' aenv) -> Gamma aenv

Construct the array environment index, will be used by code generation to map free array variable indices to names in the generated code.

freevar :: (Shape sh, Elt e) => Idx aenv (Array sh e) -> IntMap (Idx' aenv)

A free variable