Data.Array.Accelerate.LLVM.CodeGen.Exp

Description

Documentation

class Expression arch where

A class covering code generation for a subset of the scalar operations. All operations (except Foreign) have a default instance, but this allows a backend to specialise the implementation for the more complex operations.

Minimal complete definition

Nothing

Methods

eforeign :: (Foreign f, Elt x, Elt y) => arch -> f x y -> IRFun1 arch () (x -> y) -> IROpenExp arch env aenv x -> IROpenExp arch env aenv y

while :: Elt a => IROpenFun1 arch env aenv (a -> Bool) -> IROpenFun1 arch env aenv (a -> a) -> IROpenExp arch env aenv a -> IROpenExp arch env aenv a

llvmOfFun1 :: Expression arch => arch -> DelayedFun aenv (a -> b) -> Gamma aenv -> IRFun1 arch aenv (a -> b)

llvmOfFun2 :: Expression arch => arch -> DelayedFun aenv (a -> b -> c) -> Gamma aenv -> IRFun2 arch aenv (a -> b -> c)

llvmOfOpenExp :: forall arch env aenv _t. Expression arch => arch -> DelayedOpenExp env aenv _t -> Val env -> Gamma aenv -> IROpenExp arch env aenv _t

Convert an open scalar expression into a sequence of LLVM IR instructions. Code is generated in depth first order, and uses a monad to collect the sequence of instructions used to construct basic blocks.

intOfIndex :: forall sh. Shape sh => IR sh -> IR sh -> CodeGen (IR Int)

Convert a multidimensional array index into a linear index

indexOfInt :: forall sh. Shape sh => IR sh -> IR Int -> CodeGen (IR sh)

Convert a linear index into into a multidimensional index

llvmOfPrimFun :: (Elt a, Elt r) => PrimFun (a -> r) -> IR a -> CodeGen (IR r)

Generate llvm operations for primitive scalar functions