Data.Array.Accelerate.LLVM.CodeGen.IR

Description

Documentation

data IR t where

The datatype IR represents the LLVM IR producing a value of type a. Note that the operands comprising this value are stored in representation type.

Constructors

IR :: Operands (EltRepr t) -> IR t

data family Operands e :: *

class IROP dict where

Given some evidence that 'IR a' represents a scalar type, it can be converted between the IR and Operand data types.

Methods

op :: dict a -> IR a -> Operand a

ir :: dict a -> Operand a -> IR a

ir' :: dict a -> Operand a -> Operands a

op' :: dict a -> Operands a -> Operand a