Data.Array.Accelerate.LLVM.CodeGen.Loop

Description

Documentation

for

Arguments

:: (Elt i, IsIntegral i)

=> IR i
starting index

-> (IR i -> CodeGen (IR Bool))
loop test to keep going

-> (IR i -> CodeGen (IR i))
increment loop counter

-> (IR i -> CodeGen ())
body of the loop

-> CodeGen ()

TODO: Iterate over a multidimensional index space.

Build nested loops that iterate over a hype-rectangular index space between the given coordinates. The LLVM optimiser will be able to vectorise nested loops, including when we insert conversions to the corresponding linear index (e.g., in order to index arrays).

iterate :: Shape sh => IR sh -- ^ starting index -> IR sh -- ^ final index -> (IR sh -> CodeGen (IR a)) -- ^ body of the loop -> CodeGen (IR a) iterate from to body = error "CodeGen.Loop.iterate"

A standard for loop.

iter

Arguments

:: (Elt i, IsIntegral i, Elt a)

=> IR i
starting index

-> IR a
initial value

-> (IR i -> CodeGen (IR Bool))
index test to keep looping

-> (IR i -> CodeGen (IR i))
increment loop counter

-> (IR i -> IR a -> CodeGen (IR a))
loop body

-> CodeGen (IR a)

An loop with iteration count and accumulator.

while :: Elt a => (IR a -> CodeGen (IR Bool)) -> (IR a -> CodeGen (IR a)) -> IR a -> CodeGen (IR a)

A standard while loop