Data.Array.Accelerate.LLVM.CodeGen.Module

Description

Documentation

data Module arch aenv a

A compiled module consists of a number of global functions (kernels)

Constructors

Module

Fields

unModule :: Module

data Kernel arch aenv a

A fully-instantiated skeleton is a [collection of] kernel(s) that can be compiled by LLVM into a global function that we can execute.

Constructors

Kernel

Fields

unKernel :: Global