Data.Array.Accelerate.LLVM.CodeGen.Monad

Description

Documentation

data CodeGen a

runLLVM :: forall arch aenv a. (Target arch, Intrinsic arch) => CodeGen (IROpenAcc arch aenv a) -> Module arch aenv a

fresh :: forall a. Elt a => CodeGen (IR a)

Generate a fresh local reference

declare :: Global -> CodeGen ()

Add a global declaration to the symbol table

intrinsic :: String -> CodeGen Label

Get name of the corresponding intrinsic function implementing a given C function. If there is no mapping, the C function name is used.

data Block

newBlock :: String -> CodeGen Block

Create a new basic block, but don't yet add it to the block chain. You need to call setBlock to append it to the chain, so that subsequent instructions are added to this block.

Note: [Basic blocks]

The names of basic blocks are generated based on the base name provided to the newBlock function, as well as the current state (length) of the block stream. By not immediately adding new blocks to the stream, we have the advantage that:

Instructions are generated "in order", and are always appended to the stream. There is no need to search the stream for a block of the right name.

Blocks are named in groups, which helps readability. For example, the blocks for the then and else branches of a conditional, created at the same time, will be named similarly: 'if4.then' and 'if4.else', etc.

However, this leads to a slight awkwardness when walking the AST. Since a new naming group scheme is only applied *after* a call to setBlock, encountering (say) nested conditionals in the walk will generate logically distinct blocks that happen to have the same name. This means that instructions might be added to the wrong blocks, or the first set of blocks will be emitted empty and/or without a terminator.

setBlock :: Block -> CodeGen ()

Add this block to the block stream. Any instructions pushed onto the stream by instr and friends will now apply to this block.

beginBlock :: String -> CodeGen Block

Generate a new block and branch unconditionally to it.

createBlocks :: CodeGen [BasicBlock]

Extract the block state and construct the basic blocks that form a function body. The block stream is re-initialised, but module-level state such as the global symbol table is left intact.

instr :: Instruction a -> CodeGen (IR a)

Add an instruction to the state of the currently active block so that it is computed, and return the operand (LocalReference) that can be used to later refer to it.

instr' :: Instruction a -> CodeGen (Operand a)

do_ :: Instruction () -> CodeGen ()

Execute an unnamed instruction

return_ :: CodeGen ()

Return void from a basic block

retval_ :: Operand a -> CodeGen ()

Return a value from a basic block

br :: Block -> CodeGen Block

Unconditional branch. Return the name of the block that was branched from.

cbr :: IR Bool -> Block -> Block -> CodeGen Block

Conditional branch. Return the name of the block that was branched from.

phi :: forall a. Elt a => [(IR a, Block)] -> CodeGen (IR a)

Add a phi node to the top of the current block

phi' :: forall a. Elt a => Block -> IR a -> [(IR a, Block)] -> CodeGen (IR a)

addMetadata :: String -> [Maybe Metadata] -> CodeGen ()

Insert a metadata key/value pair into the current module.