Data.Array.Accelerate.LLVM.CodeGen.Skeleton

Description

Documentation

class Skeleton arch where

A class covering code generation for all of the primitive array operations. Client backends implement an instance of this class.

Minimal complete definition: * generate * fold, fold1, foldSeg, fold1Seg * scanl, scanl', scanl1, scanr, scanr', scanr1 * permute

Minimal complete definition

generate, fold, fold1, foldSeg, fold1Seg, scanl, scanl', scanl1, scanr, scanr', scanr1, permute

Methods

generate :: (Shape sh, Elt e) => arch -> Gamma aenv -> IRFun1 arch aenv (sh -> e) -> CodeGen (IROpenAcc arch aenv (Array sh e))

transform :: (Shape sh, Shape sh', Elt a, Elt b) => arch -> Gamma aenv -> IRFun1 arch aenv (sh' -> sh) -> IRFun1 arch aenv (a -> b) -> IRDelayed arch aenv (Array sh a) -> CodeGen (IROpenAcc arch aenv (Array sh' b))

map :: (Shape sh, Elt a, Elt b) => arch -> Gamma aenv -> IRFun1 arch aenv (a -> b) -> IRDelayed arch aenv (Array sh a) -> CodeGen (IROpenAcc arch aenv (Array sh b))

fold :: (Shape sh, Elt e) => arch -> Gamma aenv -> IRFun2 arch aenv (e -> e -> e) -> IRExp arch aenv e -> IRDelayed arch aenv (Array (sh :. Int) e) -> CodeGen (IROpenAcc arch aenv (Array sh e))

fold1 :: (Shape sh, Elt e) => arch -> Gamma aenv -> IRFun2 arch aenv (e -> e -> e) -> IRDelayed arch aenv (Array (sh :. Int) e) -> CodeGen (IROpenAcc arch aenv (Array sh e))

foldSeg :: (Shape sh, Elt e, Elt i, IsIntegral i) => arch -> Gamma aenv -> IRFun2 arch aenv (e -> e -> e) -> IRExp arch aenv e -> IRDelayed arch aenv (Array (sh :. Int) e) -> IRDelayed arch aenv (Segments i) -> CodeGen (IROpenAcc arch aenv (Array (sh :. Int) e))

fold1Seg :: (Shape sh, Elt e, Elt i, IsIntegral i) => arch -> Gamma aenv -> IRFun2 arch aenv (e -> e -> e) -> IRDelayed arch aenv (Array (sh :. Int) e) -> IRDelayed arch aenv (Segments i) -> CodeGen (IROpenAcc arch aenv (Array (sh :. Int) e))

scanl :: Elt e => arch -> Gamma aenv -> IRFun2 arch aenv (e -> e -> e) -> IRExp arch aenv e -> IRDelayed arch aenv (Vector e) -> CodeGen (IROpenAcc arch aenv (Vector e))

scanl' :: Elt e => arch -> Gamma aenv -> IRFun2 arch aenv (e -> e -> e) -> IRExp arch aenv e -> IRDelayed arch aenv (Vector e) -> CodeGen (IROpenAcc arch aenv (Vector e, Scalar e))

scanl1 :: Elt e => arch -> Gamma aenv -> IRFun2 arch aenv (e -> e -> e) -> IRDelayed arch aenv (Vector e) -> CodeGen (IROpenAcc arch aenv (Vector e))

scanr :: Elt e => arch -> Gamma aenv -> IRFun2 arch aenv (e -> e -> e) -> IRExp arch aenv e -> IRDelayed arch aenv (Vector e) -> CodeGen (IROpenAcc arch aenv (Vector e))

scanr' :: Elt e => arch -> Gamma aenv -> IRFun2 arch aenv (e -> e -> e) -> IRExp arch aenv e -> IRDelayed arch aenv (Vector e) -> CodeGen (IROpenAcc arch aenv (Vector e, Scalar e))

scanr1 :: Elt e => arch -> Gamma aenv -> IRFun2 arch aenv (e -> e -> e) -> IRDelayed arch aenv (Vector e) -> CodeGen (IROpenAcc arch aenv (Vector e))

permute :: (Shape sh, Shape sh', Elt e) => arch -> Gamma aenv -> IRFun2 arch aenv (e -> e -> e) -> IRFun1 arch aenv (sh -> sh') -> IRDelayed arch aenv (Array sh e) -> CodeGen (IROpenAcc arch aenv (Array sh' e))

backpermute :: (Shape sh, Shape sh', Elt e) => arch -> Gamma aenv -> IRFun1 arch aenv (sh' -> sh) -> IRDelayed arch aenv (Array sh e) -> CodeGen (IROpenAcc arch aenv (Array sh' e))

stencil :: (Elt a, Elt b, Stencil sh a stencil) => arch -> Gamma aenv -> IRFun1 arch aenv (stencil -> b) -> Boundary (IR a) -> IRManifest arch aenv (Array sh a) -> CodeGen (IROpenAcc arch aenv (Array sh b))

stencil2 :: (Elt a, Elt b, Elt c, Stencil sh a stencil1, Stencil sh b stencil2) => arch -> Gamma aenv -> IRFun2 arch aenv (stencil1 -> stencil2 -> c) -> Boundary (IR a) -> IRManifest arch aenv (Array sh a) -> Boundary (IR b) -> IRManifest arch aenv (Array sh b) -> CodeGen (IROpenAcc arch aenv (Array sh c))