Data.Array.Accelerate.LLVM.CodeGen.Sugar

Description

Documentation

type IRExp arch aenv t = IROpenExp arch () aenv t

LLVM IR is in single static assignment, so we need to be able to generate fresh names for each application of a scalar function or expression.

type IRFun1 arch aenv t = IROpenFun1 arch () aenv t

type IRFun2 arch aenv t = IROpenFun2 arch () aenv t

type IROpenExp arch env aenv t = CodeGen (IR t)

data IROpenFun1 arch env aenv t where

Constructors

IRFun1 :: (IR a -> IROpenExp arch (env, a) aenv b) -> IROpenFun1 arch env aenv (a -> b)

Fields

app1 :: IR a -> IROpenExp arch (env, a) aenv b

data IROpenFun2 arch env aenv t where

Constructors

IRFun2 :: (IR a -> IR b -> IROpenExp arch ((env, a), b) aenv c) -> IROpenFun2 arch env aenv (a -> b -> c)

Fields

app2 :: IR a -> IR b -> IROpenExp arch ((env, a), b) aenv c

data IROpenAcc arch aenv arrs where

Constructors

IROpenAcc :: [Kernel arch aenv a] -> IROpenAcc arch aenv arrs

data IRDelayed arch aenv a where

Constructors

IRDelayed :: (Shape sh, Elt e) => IRExp arch aenv sh -> IRFun1 arch aenv (sh -> e) -> IRFun1 arch aenv (Int -> e) -> IRDelayed arch aenv (Array sh e)

Fields

delayedExtent :: IRExp arch aenv sh

delayedIndex :: IRFun1 arch aenv (sh -> e)

delayedLinearIndex :: IRFun1 arch aenv (Int -> e)

data IRManifest arch aenv a where

Constructors

IRManifest :: Arrays arrs => Idx aenv arrs -> IRManifest arch aenv arrs

data IRArray a where

Constructors

IRArray :: (Shape sh, Elt e) => IR sh -> IR e -> IRArray (Array sh e)

Fields

irArrayShape :: IR sh

irArrayData :: IR e