Data.Array.Accelerate.LLVM.CodeGen.Type

Description

Documentation

class IsSigned dict where

Does the concrete type represent signed or unsigned values?

Minimal complete definition

Nothing

Methods

signed :: dict a -> Bool

unsigned :: dict a -> Bool

class TypeOf op where

Extract the reified scalar type dictionary of an operation

Methods

typeOf :: op a -> ScalarType a