Data.Array.Accelerate.LLVM.Compile

Description

Documentation

class Compile arch where

Associated Types

data ExecutableR arch

Methods

compileForTarget :: DelayedOpenAcc aenv a -> Gamma aenv -> LLVM arch (ExecutableR arch)

Compile an accelerate computation into some backend-specific executable format

compileAcc :: (Compile arch, Remote arch) => DelayedAcc a -> LLVM arch (ExecAcc arch a)

Initialise code generation, compilation, and data transfer (if required) for an array expression. The returned array computation is annotated to be suitable for execution on the target:

A list of the array variables embedded within scalar expressions

The compiled LLVM code required to execute the kernel

compileAfun :: (Compile arch, Remote arch) => DelayedAfun f -> LLVM arch (ExecAfun arch f)

data ExecOpenAcc arch aenv a where

Annotate an open array expression with the information necessary to execute each node directly.

Constructors

ExecAcc :: ExecutableR arch -> Gamma aenv -> PreOpenAcc (ExecOpenAcc arch) aenv a -> ExecOpenAcc arch aenv a

EmbedAcc :: (Shape sh, Elt e) => PreExp (ExecOpenAcc arch) aenv sh -> ExecOpenAcc arch aenv (Array sh e)

type ExecAcc arch a = ExecOpenAcc arch () a

type ExecAfun arch a = PreAfun (ExecOpenAcc arch) a

type ExecExp arch = ExecOpenExp arch ()

type ExecOpenExp arch = PreOpenExp (ExecOpenAcc arch)

type ExecFun arch = ExecOpenFun arch ()

type ExecOpenFun arch = PreOpenFun (ExecOpenAcc arch)