Data.Array.Accelerate.LLVM.Execute

Description

Documentation

class Remote arch => Execute arch where

Methods

map :: (Shape sh, Elt b) => ExecutableR arch -> Gamma aenv -> AvalR arch aenv -> StreamR arch -> sh -> LLVM arch (Array sh b)

generate :: (Shape sh, Elt e) => ExecutableR arch -> Gamma aenv -> AvalR arch aenv -> StreamR arch -> sh -> LLVM arch (Array sh e)

transform :: (Shape sh, Elt e) => ExecutableR arch -> Gamma aenv -> AvalR arch aenv -> StreamR arch -> sh -> LLVM arch (Array sh e)

backpermute :: (Shape sh, Elt e) => ExecutableR arch -> Gamma aenv -> AvalR arch aenv -> StreamR arch -> sh -> LLVM arch (Array sh e)

fold :: (Shape sh, Elt e) => ExecutableR arch -> Gamma aenv -> AvalR arch aenv -> StreamR arch -> (sh :. Int) -> LLVM arch (Array sh e)

fold1 :: (Shape sh, Elt e) => ExecutableR arch -> Gamma aenv -> AvalR arch aenv -> StreamR arch -> (sh :. Int) -> LLVM arch (Array sh e)

foldSeg :: (Shape sh, Elt e) => ExecutableR arch -> Gamma aenv -> AvalR arch aenv -> StreamR arch -> (sh :. Int) -> DIM1 -> LLVM arch (Array (sh :. Int) e)

fold1Seg :: (Shape sh, Elt e) => ExecutableR arch -> Gamma aenv -> AvalR arch aenv -> StreamR arch -> (sh :. Int) -> DIM1 -> LLVM arch (Array (sh :. Int) e)

scanl :: Elt e => ExecutableR arch -> Gamma aenv -> AvalR arch aenv -> StreamR arch -> DIM1 -> LLVM arch (Vector e)

scanl1 :: Elt e => ExecutableR arch -> Gamma aenv -> AvalR arch aenv -> StreamR arch -> DIM1 -> LLVM arch (Vector e)

scanl' :: Elt e => ExecutableR arch -> Gamma aenv -> AvalR arch aenv -> StreamR arch -> DIM1 -> LLVM arch (Vector e, Scalar e)

scanr :: Elt e => ExecutableR arch -> Gamma aenv -> AvalR arch aenv -> StreamR arch -> DIM1 -> LLVM arch (Vector e)

scanr1 :: Elt e => ExecutableR arch -> Gamma aenv -> AvalR arch aenv -> StreamR arch -> DIM1 -> LLVM arch (Vector e)

scanr' :: Elt e => ExecutableR arch -> Gamma aenv -> AvalR arch aenv -> StreamR arch -> DIM1 -> LLVM arch (Vector e, Scalar e)

permute :: (Shape sh, Shape sh', Elt e) => ExecutableR arch -> Gamma aenv -> AvalR arch aenv -> StreamR arch -> sh -> Array sh' e -> LLVM arch (Array sh' e)

stencil1 :: (Shape sh, Elt a, Elt b) => ExecutableR arch -> Gamma aenv -> AvalR arch aenv -> StreamR arch -> Array sh a -> LLVM arch (Array sh b)

stencil2 :: (Shape sh, Elt a, Elt b, Elt c) => ExecutableR arch -> Gamma aenv -> AvalR arch aenv -> StreamR arch -> Array sh a -> Array sh b -> LLVM arch (Array sh c)

aforeign :: (Arrays as, Arrays bs, Foreign f) => f as bs -> ExecAfun arch (as -> bs) -> as -> LLVM arch bs

type Gamma aenv = IntMap (Label, Idx' aenv)

A mapping between the environment index of a free array variable and the Name of that array to be used in the generated code.

This simply compresses the array indices into a continuous range, rather than directly using the integer equivalent of the de Bruijn index. Thus, the result is still sensitive to the order of let bindings, but not of any intermediate (unused) free array variables.

executeAcc :: forall arch a. (Execute arch, Arrays a) => ExecAcc arch a -> LLVM arch a

executeAfun1 :: forall arch a b. (Execute arch, Arrays a, Arrays b) => ExecAfun arch (a -> b) -> a -> LLVM arch b