Data.Array.Accelerate.LLVM.Execute.Async

Description

Documentation

data AsyncR arch a

The result of a potentially parallel computation which will be available at some point (presumably, in the future). This is essentially a write-once IVar.

Constructors

AsyncR !(EventR arch) !a

class Async arch where

Associated Types

type StreamR arch

Streams (i.e. threads) can execute concurrently with other streams, but operations within the same stream proceed sequentially.

type EventR arch

An Event marks a point in the execution stream, possibly in the future. Since execution within a stream is sequential, events can be used to test the progress of a computation and synchronise between different streams.

Methods

spawn :: LLVM arch (StreamR arch)

Create a new execution stream that can be used to track (potentially parallel) computations

checkpoint :: StreamR arch -> LLVM arch (EventR arch)

Generate a new event at the end of the given execution stream. It will be filled once all prior work submitted to the stream has completed.

after :: StreamR arch -> EventR arch -> LLVM arch ()

Make all future work submitted to the given execution stream wait until the given event has passed. Typically the event is from a different execution stream, therefore this function is intended to enable non-blocking cross-stream coordination.

block :: EventR arch -> LLVM arch ()

Block execution of the calling thread until the given event has been recorded.

get :: Async arch => AsyncR arch a -> LLVM arch a

Wait for an asynchronous operation to complete, then return it.

async :: Async arch => (StreamR arch -> LLVM arch a) -> LLVM arch (AsyncR arch a)

Execute the given operation asynchronously in a new execution stream.