Data.Array.Accelerate.LLVM.Execute.Marshal

Description

Documentation

marshal :: Marshalable t args => t -> StreamR t -> args -> IO [ArgR t]

Convert function arguments into stream a form suitable for CUDA function calls

type family ArgR target

A type family that is used to specify a concrete kernel argument and stream/context type for a given backend target.

class Marshalable t a where

Data which can be marshalled as function arguments to kernels.

These are just the basic definitions that don't require backend specific knowledge. To complete the definition, a backend must provide instances for:

Int -- for shapes

ArrayData e -- for array data

(Gamma aenv, Aval aenv) -- for free array variables

Methods

marshal' :: t -> StreamR t -> a -> IO (DList (ArgR t))