Data.Array.Accelerate.LLVM.State

Description

Documentation

newtype LLVM target a

The LLVM monad, for executing array computations. This consists of a stack for the LLVM execution context as well as the per-execution target specific state target.

Constructors

LLVM

Fields

runLLVM :: StateT target IO a

llvmTarget :: t -> t

Extract the execution state: 'gets llvmTarget'

evalLLVM :: t -> LLVM t a -> IO a

Evaluate the given target with an LLVM context

keepAlive :: a -> IO a

Make sure the GC knows that we want to keep this thing alive forever.

We may want to introduce some way to actually shut this down if, for example, the object has not been accessed in a while (whatever that means).

Broken in ghci-7.6.1 Mac OS X due to bug #7299.