Data.Array.Accelerate.LLVM.Util

Description

Generate types for the reified elements of an array computation

Documentation

bitSize :: Bits a => a -> Word32

The number of bits in a type

fromBool :: Integral i => Bool -> i

Convert a boolean value into an integral value, where False is zero and True is one.

toBool :: Integral i => i -> Bool

Convert an integral value into a boolean. We follow the C convention, where zero is False and all other values represent True.