Data.IORef.Storable

Description

A mutable reference in the IO monad.

Documentation

data IORef a

A pointer to a mutable object. Unlike IORef, this is just a pointer to the value, rather than a reference containing a boxed object. This means that the value is stored strictly, and is only one pointer away.

newIORef :: Storable a => a -> IO (IORef a)

Build a new IORef

readIORef :: Storable a => IORef a -> IO a

Read the value of the IORef

writeIORef :: Storable a => IORef a -> a -> IO ()

Write a new value into the IORef

modifyIORef :: Storable a => IORef a -> (a -> a) -> IO ()

Mutate the contents of an IORef