Data.Range.Range

Description

Documentation

data Range

A simple range data type

Constructors

Empty
The empty range

IE !Int !Int
A range span with inclusive left, exclusive right

empty :: Range

An empty interval

null :: Range -> Bool

Check if an interval is empty

singleton :: Int -> Range

A singleton point

(...) :: Int -> Int -> Range infix 3

A range span with exclusive endpoint [u,v).

size :: Range -> Int

O(1). The number of elements defined by the range interval

bisect :: Range -> (Range, Range)

O(1). Split an interval into two roughly equally sized ranges. If the interval is odd then the first interval gets the extra element.

take :: Int -> Range -> Range

O(1). Return the first n elements of the range, or the range itself if n > size.

splitAt :: Int -> Range -> (Range, Range)

O(1). A tuple where the first element is the first n elements of the range, and the second is the remainder of the list (if any).

merge :: Range -> Range -> Maybe Range

If the two ranges are adjacent, return one combined range. The ranges must not be empty.

append :: Seq Range -> Range -> Seq Range

O(1). Add a new range to the end of the given sequence. We assume that ranges are non-overlapping and non-empty. If the new range is adjacent to the last range on the sequence, the ranges are appended.

compress :: Seq Range -> Seq Range

O(n log n). Compress the given ranges into the fewest number of sections as possible. The ranges must not be empty.