LLVM.General.AST.Type.Constant

Description

Documentation

data Constant a where

Although constant expressions and instructions have many similarities, there are important differences - so they're represented using different types in this AST. At the cost of making it harder to move an code back and forth between being constant and not, this approach embeds more of the rules of what IR is legal into the Haskell types.

http://llvm.org/docs/LangRef.html#constants

http://llvm.org/docs/LangRef.html#constant-expressions

Constructors

ScalarConstant :: ScalarType a -> a -> Constant a

GlobalReference :: Maybe (ScalarType a) -> Name a -> Constant a