LLVM.General.AST.Type.Global

Description

Documentation

data Parameter a where

Parameters for functions

Constructors

ScalarParameter :: ScalarType a -> Name a -> Parameter a

PtrParameter :: ScalarType a -> Name a -> Parameter (Ptr a)

data FunctionAttribute

Attributes for the function call instruction

Constructors

NoReturn

NoUnwind

ReadOnly

ReadNone

AlwaysInline

data GroupID

Attribute groups are groups of attributes that are referenced by objects within the IR. To use an attribute group, an object must reference its GroupID.

Constructors

GroupID !Word

data GlobalFunction args t where

A global function definition

Note that because we just use the reified dictionary structure of Accelerate types, our functions are limited to operating over scalar types only; no pointers to functions and nothing that returns void.

Constructors

Body :: Maybe (ScalarType r) -> Label -> GlobalFunction `[]` r

Lam :: ScalarType a -> Operand a -> GlobalFunction args t -> GlobalFunction (a : args) t

data HList l where

Constructors

HNil :: HList `[]`

HCons :: e -> HList l -> HList (e : l)