LLVM.General.AST.Type.Terminator

Description

Documentation

data Terminator a where

http://llvm.org/docs/LangRef.html#terminators

TLM: well, I don't think the types of these terminators make any sense. When we branch, we are not propagating a particular value, just moving the program counter, and anything we have declared already is available for later computations. Maybe, we can make some of this explicit in the phi node?

Constructors

Ret :: Terminator ()

RetVal :: Operand a -> Terminator a

Br :: Label -> Terminator ()

CondBr :: Operand Bool -> Label -> Label -> Terminator ()

Switch :: Operand a -> Label -> [(Constant a, Label)] -> Terminator ()