Data.Array.Accelerate.LLVM.CodeGen.Downcast

Description

Documentation

class Downcast typed untyped where

Convert a value from our representation of the LLVM AST which uses Haskell-level types, into the llvm-general representation where types are represented only at the value level. We use the type-level information to generate the appropriate value-level types.

Methods

downcast :: typed -> untyped